Hiang Nguyen up code ver 3
